Add deleteEventFromNamedCalendarWithID function to plugin.
Was not able to get the other delete functions to work, but
created this function to delete by id. For some reason we
need to reconnect to the event store when deleting, if it
is a newly created event, or else ALL instances of the
event will be deleted no matter what. I am not sure why
this is.